KEM might require other changes.
As pointed out be Meiling Cheng, the introduction of a KEM might require other changes to the protocol. Meiling Cheng pointed out that using a KEM instead of ECDHE means that the nice property that both parties contribute randomness to the session key is lost. This could be fixed by including the ephemeral public key of the server in the key derivation. Other changes might be required as well.

#43